Rollback changes for non-swing interop
