--- skiko/src/awtMain/kotlin/org/jetbrains/skiko/SkiaDrawingManager.kt
@@ -1,126 +1,0 @@
-package org.jetbrains.skiko
-
-import org.jetbrains.skia.*
-import javax.swing.SwingUtilities
-
-internal class SkiaDrawingManager(
-    private val fpsCounter: FPSCounter?,
-    private val clipComponents: MutableList<ClipRectangle>
-) {
-    private var _isRendering = false
-    val isRendering: Boolean
-        get() = _isRendering
-
-    @Volatile
-    private var isDisposed = false
-
-    @Volatile
-    private var picture: PictureHolder? = null
-    private var pictureRecorder: PictureRecorder? = null
-    private val pictureLock = Any()
-
-    fun init() {
-        isDisposed = false
-        pictureRecorder = PictureRecorder()
-    }
-
-    fun dispose() {
-        picture?.instance?.close()
-        picture = null
-        pictureRecorder?.close()
-        pictureRecorder = null
-        isDisposed = true
-    }
-
-    fun update(
-        nanoTime: Long,
-        width: Int,
-        height: Int,
-        contentScale: Float,
-        skikoView: SkikoView?
-    ) {
-        check(SwingUtilities.isEventDispatchThread()) { "Method should be called from AWT event dispatch thread" }
-
-        FrameWatcher.nextFrame()
-        fpsCounter?.tick()
-
-        val pictureWidth = (width * contentScale).toInt().coerceAtLeast(0)
-        val pictureHeight = (height * contentScale).toInt().coerceAtLeast(0)
-
-        val bounds = Rect.makeWH(pictureWidth.toFloat(), pictureHeight.toFloat())
-        val pictureRecorder = pictureRecorder!!
-        val canvas = pictureRecorder.beginRecording(bounds)
-
-        // clipping
-        for (component in clipComponents) {
-            canvas.clipRectBy(component, contentScale)
-        }
-
-        try {
-            _isRendering = true
-            skikoView?.onRender(canvas, pictureWidth, pictureHeight, nanoTime)
-        } finally {
-            _isRendering = false
-        }
-
-        // we can dispose layer during onRender
-        // or even dispose it and pack it again
-        if (!isDisposed && !pictureRecorder.isClosed) {
-            synchronized(pictureLock) {
-                picture?.instance?.close()
-                val picture = pictureRecorder.finishRecordingAsPicture()
-                this.picture = PictureHolder(picture, pictureWidth, pictureHeight)
-            }
-        }
-    }
-
-    fun draw(canvas: Canvas) {
-        check(!isDisposed) { "SkiaLayer is disposed" }
-        lockPicture {
-            canvas.drawPicture(it.instance)
-        }
-    }
-
-    // Captures current layer as bitmap.
-    fun screenshot(): Bitmap? {
-        check(!isDisposed) { "SkiaLayer is disposed" }
-        return lockPicture { picture ->
-            val store = Bitmap()
-            val ci = ColorInfo(
-                ColorType.BGRA_8888, ColorAlphaType.OPAQUE, ColorSpace.sRGB
-            )
-            store.setImageInfo(ImageInfo(ci, picture.width, picture.height))
-            store.allocN32Pixels(picture.width, picture.height)
-            val canvas = Canvas(store)
-            canvas.drawPicture(picture.instance)
-            store.setImmutable()
-            store
-        }
-    }
-
-
-    private fun <T : Any> lockPicture(action: (PictureHolder) -> T): T? {
-        return synchronized(pictureLock) {
-            val picture = picture
-            if (picture != null) {
-                action(picture)
-            } else {
-                null
-            }
-        }
-    }
-}
-
-internal fun Canvas.clipRectBy(rectangle: ClipRectangle, contentScale: Float) {
-    val dpi = contentScale
-    clipRect(
-        Rect.makeLTRB(
-            rectangle.x * dpi,
-            rectangle.y * dpi,
-            (rectangle.x + rectangle.width) * dpi,
-            (rectangle.y + rectangle.height) * dpi
-        ),
-        ClipMode.DIFFERENCE,
-        true
-    )
-}
