+++ skiko/src/awtMain/kotlin/org/jetbrains/skiko/context/JvmContextHandler.kt
@@ -2,4 +2,4 @@ package org.jetbrains.skiko.context
 
 import org.jetbrains.skiko.SkiaLayer
 
-internal abstract class JvmContextHandler(layer: SkiaLayer) : SkiaLayerContextHandler(layer, layer::draw)
+internal abstract class JvmContextHandler(layer: SkiaLayer) : ContextHandler(layer, layer::draw)
