+++ skiko/src/awtMain/kotlin/org/jetbrains/skiko/context/MetalContextHandler.kt
@@ -2,7 +2,6 @@ package org.jetbrains.skiko.context
 
 import org.jetbrains.skia.*
 import org.jetbrains.skiko.Logger
-import org.jetbrains.skiko.MetalAdapter
 import org.jetbrains.skiko.RenderException
 import org.jetbrains.skiko.SkiaLayer
 import org.jetbrains.skiko.redrawer.MetalDevice
@@ -18,7 +17,7 @@ import org.jetbrains.skiko.redrawer.MetalDevice
 internal class MetalContextHandler(
     layer: SkiaLayer,
     private val device: MetalDevice,
-    private val adapter: MetalAdapter
+    private val adapterInfo: AdapterInfo? = null
 ) : JvmContextHandler(layer) {
     override fun initContext(): Boolean {
         try {
@@ -70,8 +69,11 @@ internal class MetalContextHandler(
 
     override fun rendererInfo(): String {
         return super.rendererInfo() +
-                "Video card: ${adapter.name}\n" +
-                "Total VRAM: ${adapter.memorySize / 1024 / 1024} MB\n"
+                if (adapterInfo != null) {
+                    "Video card: ${adapterInfo.adapterName}\n" + "Total VRAM: ${adapterInfo.adapterMemorySize / 1024 / 1024} MB\n"
+                } else {
+                    ""
+                }
     }
 
     private fun makeRenderTarget(width: Int, height: Int) = BackendRenderTarget(
@@ -87,4 +89,6 @@ internal class MetalContextHandler(
     private external fun makeMetalContext(device: Long): Long
     private external fun makeMetalRenderTarget(device: Long, width: Int, height: Int): Long
     private external fun finishFrame(device: Long)
+
+    data class AdapterInfo(val adapterName: String, val adapterMemorySize: Long)
 }
