+++ skiko/src/commonMain/kotlin/org/jetbrains/skiko/context/ContextHandler.kt
@@ -1,17 +1,12 @@
 package org.jetbrains.skiko.context
 
 import org.jetbrains.skia.*
-import org.jetbrains.skiko.GraphicsApi
-import org.jetbrains.skiko.RenderException
-import org.jetbrains.skiko.hostArch
-import org.jetbrains.skiko.hostOs
+import org.jetbrains.skiko.*
 
 internal abstract class ContextHandler(
+    protected val layer: SkiaLayer,
     private val drawContent: Canvas.() -> Unit
 ) {
-    abstract val renderApi: GraphicsApi
-    protected abstract fun isTransparentBackground(): Boolean
-
     protected var context: DirectContext? = null
     protected var renderTarget: BackendRenderTarget? = null
     protected var surface: Surface? = null
@@ -35,7 +30,7 @@ internal abstract class ContextHandler(
     }
 
     open fun rendererInfo(): String {
-        return "GraphicsApi: ${renderApi}\n" +
+        return "GraphicsApi: ${layer.renderApi}\n" +
                 "OS: ${hostOs.id} ${hostArch.id}\n"
     }
 
@@ -50,5 +45,18 @@ internal abstract class ContextHandler(
             drawContent()
         }
         flush()
+    }
+
+    protected open fun isTransparentBackground(): Boolean {
+        if (hostOs == OS.MacOS) {
+            // MacOS transparency is always supported
+            return true
+        }
+        if (layer.fullscreen) {
+            // for non-MacOS in fullscreen transparency is not supported
+            return false
+        }
+        // for non-MacOS in non-fullscreen transparency provided by [layer]
+        return layer.transparency
     }
 }
