+++ skiko/src/iosMain/kotlin/org/jetbrains/skiko/context/MetalContextHandler.kt
@@ -6,7 +6,7 @@ import org.jetbrains.skiko.RenderException
 import org.jetbrains.skiko.SkiaLayer
 import org.jetbrains.skiko.redrawer.MetalRedrawer
 
-internal class MetalContextHandler(layer: SkiaLayer) : SkiaLayerContextHandler(layer, layer::draw) {
+internal class MetalContextHandler(layer: SkiaLayer) : ContextHandler(layer, layer::draw) {
     val metalRedrawer: MetalRedrawer
         get() = layer.redrawer!!
